Swap Home and Library fragment@
