+++ app/src/main/java/com/example/dinhh/soundscape/presentation/screens/library/LibararyAdapter.kt
@@ -1,4 +1,4 @@
-package com.example.dinhh.soundscape.presentation.screens.home
+package com.example.dinhh.soundscape.presentation.screens.library
 
 import android.content.Context
 import android.support.v7.widget.RecyclerView
