+++ app/src/main/java/com/example/dinhh/soundscape/presentation/screens/library/LibraryFragment.kt
@@ -1,55 +1,67 @@
 package com.example.dinhh.soundscape.presentation.screens.library
 
+
+import android.content.Intent
 import android.os.Bundle
 import android.support.v4.app.Fragment
+import android.support.v7.widget.LinearLayoutManager
+import android.support.v7.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import com.example.dinhh.soundscape.R
-import com.example.dinhh.soundscape.presentation.screens.sounds.SoundFragment
-import kotlinx.android.synthetic.main.fragment_library.*
-
+import com.example.dinhh.soundscape.presentation.base.RecyclerViewListener
+import com.example.dinhh.soundscape.presentation.screens.savedrecord.SavedRecord
 
 class LibraryFragment : Fragment() {
+
+    private lateinit var adapter: LibararyAdapter
+    private lateinit var rvLibrary: RecyclerView
+
+    private lateinit var myLibraryTitle: List<String>
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        return inflater.inflate(R.layout.fragment_library, container, false)
+
+        // Inflate the layout for this fragment
+        val view = inflater.inflate(R.layout.fragment_library, container, false)
+        rvLibrary = view.findViewById(R.id.rvLibrary) as RecyclerView
+
+        myLibraryTitle = listOf(
+            getString(R.string.save_soundscapes),
+            getString(R.string.recorded_tracks),
+            getString(R.string.favorite_sounds)
+        )
+
+        setupViews()
+
+        return view
     }
 
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-        setupButtons()
+    private fun setupViews() {
+        adapter = LibararyAdapter(myLibraryTitle)
+        adapter.setOnItemClickListener(object : RecyclerViewListener.OnItemClickListener {
+            override fun onItemClick(view: View, position: Int) {
+                when (position) {
+                    0 -> {
+                        goToSavedSoundscapes()
+                    }
+                }
+            }
+        })
+        val layoutManager = LinearLayoutManager(context!!)
+        rvLibrary.layoutManager = layoutManager
+        rvLibrary.adapter = adapter
     }
 
-    private fun setupButtons(){
-        natureBtn.setOnClickListener {
-            goToSoundsFragment("nature")
-        }
-        humanBtn.setOnClickListener {
-            goToSoundsFragment("human")
-        }
-        machineBtn.setOnClickListener {
-            goToSoundsFragment("machine")
-        }
-        storyBtn.setOnClickListener {
-            goToSoundsFragment("story")
-        }
-
+    private fun goToSavedSoundscapes() {
+        val intent = Intent(activity, SavedRecord::class.java)
+        startActivity(intent)
     }
-
-    private fun goToSoundsFragment(category: String){
-        val fragManager = fragmentManager
-        val fragmentTransaction = fragManager?.beginTransaction()
-        fragmentTransaction?.replace(R.id.container, SoundFragment.newInstance(category))
-        fragmentTransaction?.addToBackStack(null)?.commit()
-    }
-
     companion object {
         @JvmStatic
         fun newInstance() = LibraryFragment()
     }
-
 }
